requirements.txt
# OCR Engines
paddleocr==2.7.3
easyocr==1.7.1
pytesseract==0.3.10

# PDF Processing
pdf2image==1.17.0
PyMuPDF==1.23.26
pikepdf==8.10.1

# Image Processing
opencv-python==4.9.0.80
Pillow==10.2.0
scikit-image==0.22.0

# Deep Learning
torch==2.2.0
torchvision==0.17.0
transformers==4.37.2

# Google Drive Integration
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0
pydrive2==1.18.1

# Data Processing
numpy==1.26.3
pandas==2.2.0

# Multiprocessing & Queue
redis==5.0.1
celery==5.3.6

# Database
sqlalchemy==2.0.25

# Progress & Logging
tqdm==4.66.1
loguru==0.7.2
rich==13.7.0

# Arabic Text Processing
python-bidi==0.4.2
arabic-reshaper==3.0.0
pyarabic==0.6.15

# Configuration
python-dotenv==1.0.1
pyyaml==6.0.1

# Output Formats
python-docx==1.1.0
reportlab==4.0.9
openpyxl==3.1.2

# Utils
requests==2.31.0
